Removed app.use(express.json)
This is actually very weird, but it was causing a bug which I could not
identify. During testing, making a GET request to
localhost:4000/api/room/hi (or any room name) would just cause the page
to continually load. Unfortunately, since I could not identify why that
line caused the issue, I just decided to comment it out.